refactor: added dropdownInput to schema
