--- types/dropdown.ts
@@ -1,4 +1,0 @@
-export interface DropdownOption {
-  label: string;
-  value: string;
-}
